Move data sync to a Cloudflare Pages Function (no Worker deploy needed)
Extracts the Atlas Academy -> Supabase pipeline into shared/atlasSync.js and
adds functions/api/sync.js, which ships with the Pages build and triggers the
sync on demand. GET returns sync status; POST triggers, gated by a shared
cooldown + JP-hash freshness check so an open endpoint can't be abused or
inject data. The standalone Worker now imports the same shared module and is
only needed for the optional cron. DataUpdateButton points at /api/sync.
