+++ src/components/DataUpdateButton.js
@@ -5,10 +5,10 @@ import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import WarningIcon from '@mui/icons-material/Warning';
 import { supabase } from '../supabaseClient';
 
-// When deployed as a single Cloudflare Worker, REACT_APP_WORKER_URL is not
-// needed — /health and /run are served by the same origin.  Set it only for
-// local development to point at `wrangler dev` (e.g. http://localhost:8787).
-const WORKER_URL = process.env.REACT_APP_WORKER_URL || '';
+// Sync runs as a Cloudflare Pages Function at /api/sync (same origin as the
+// app), so no separate Worker URL is needed. REACT_APP_WORKER_URL only matters
+// for local dev if you proxy the function from another origin.
+const SYNC_URL = `${process.env.REACT_APP_WORKER_URL || ''}/api/sync`;
 
 function fmtAge(iso) {
   if (!iso) return null;
@@ -43,7 +43,7 @@ export default function DataUpdateButton() {
   useEffect(() => { fetchLastUpdated(); }, [fetchLastUpdated]);
 
   useEffect(() => {
-    fetch(`${WORKER_URL}/health`)
+    fetch(SYNC_URL)
       .then(r => setWorkerOk(r.ok))
       .catch(() => setWorkerOk(false));
   }, []);
@@ -53,13 +53,17 @@ export default function DataUpdateButton() {
     setError(null);
     setNotice(null);
     try {
-      const res = await fetch(`${WORKER_URL}/run`, { method: 'POST' });
+      const res = await fetch(SYNC_URL, { method: 'POST' });
       const body = await res.json().catch(() => ({}));
       if (res.status === 429 || body.status === 'skipped') {
         setNotice('Data is already up to date — try again later.');
         return;
       }
-      if (!res.ok) throw new Error(`Worker ${res.status}`);
+      if (body.status === 'up_to_date') {
+        setNotice('Already on the latest game version.');
+        return;
+      }
+      if (!res.ok) throw new Error(`Sync ${res.status}`);
       setTimeout(fetchLastUpdated, 5000);
       setTimeout(fetchLastUpdated, 15000);
     } catch (e) {
